# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Test Management Platform API
  description: |
    Comprehensive API for the Test Management Platform with Observer/Orchestrator architecture.
    
    This API provides endpoints for:
    - Authentication and user management
    - Test execution orchestration
    - Git repository management
    - Webhook integrations for CI/CD systems
    
    ## Authentication
    The API uses session-based authentication. Users must login to receive a session cookie
    that will be used for subsequent authenticated requests.
    
    ## Architecture
    The platform follows an Observer/Orchestrator pattern where the API coordinates
    external CI/CD systems rather than executing tests directly.
  version: 1.0.0
  contact:
    name: Test Management Platform
    url: http://localhost:3000
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /git/repositories:
    get:
      tags:
        - Git Repositories
      summary: List git repositories
      description: Get paginated list of git repositories
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of git repositories
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      $ref: '#/components/schemas/GitRepository'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Git Repositories
      summary: Create git repository
      description: Add a new git repository to the system
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - url
              properties:
                name:
                  type: string
                  example: my-test-repo
                url:
                  type: string
                  format: uri
                  example: https://github.com/user/repo.git
                branch:
                  type: string
                  default: main
                  example: main
                description:
                  type: string
                  example: Repository for automated testing
                provider:
                  type: string
                  enum: [github, gitlab, bitbucket, azure]
                  default: github
                access_token:
                  type: string
                  description: Access token for private repositories
                webhook_secret:
                  type: string
                  description: Secret for webhook validation
      responses:
        '201':
          description: Repository created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Git repository created successfully
                  repository:
                    $ref: '#/components/schemas/GitRepository'
  /git/repositories/{id}:
    get:
      tags:
        - Git Repositories
      summary: Get git repository
      description: Get details of a specific git repository
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Git repository details
          content:
            application/json:
              schema:
                type: object
                properties:
                  repository:
                    $ref: '#/components/schemas/GitRepository'

    put:
      tags:
        - Git Repositories
      summary: Update git repository
      description: Update git repository configuration
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
                branch:
                  type: string
                description:
                  type: string
                provider:
                  type: string
                  enum: [github, gitlab, bitbucket, azure]
                access_token:
                  type: string
                webhook_secret:
                  type: string
                status:
                  type: string
                  enum: [active, inactive, error]
      responses:
        '200':
          description: Repository updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Git repository updated successfully

    delete:
      tags:
        - Git Repositories
      summary: Delete git repository
      description: Remove git repository from the system
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Repository deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Git repository deleted successfully
  /git/repositories/{id}/test-connection:
    post:
      tags:
        - Git Repositories
      summary: Test repository connection
      description: Test connection to git repository
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  repository:
                    type: object
                    properties:
                      name:
                        type: string
                      url:
                        type: string
                      branch:
                        type: string
                      status:
                        type: string
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using HTTP cookies
  schemas:
...
    GitRepository:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: my-test-repo
        url:
          type: string
          format: uri
          example: https://github.com/user/repo.git
        branch:
          type: string
          example: main
        description:
          type: string
        provider:
          type: string
          enum: [github, gitlab, bitbucket, azure]
        status:
          type: string
          enum: [active, inactive, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: integer
        created_by_username:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...